+++ Runtime/Bus/IEventListener.cs
@@ -1,7 +1,0 @@
-﻿namespace AceLand.EventDriven.Bus
-{
-    public interface IEventListener
-    {
-        // Marker Interface for Event Listener
-    }
-}