+++ Runtime/Bus/IEventRaiser.cs
@@ -1,7 +1,0 @@
-﻿namespace AceLand.EventDriven.Bus
-{
-    public interface IEventRaiser
-    {
-        // Marker Interface for Event Raiser
-    }
-}